Move childProcess externals to a subpackage
Aligns with how kotlin-wrappers does it, and the emitted import looks saner: it
immediately just gets the 'spawn' property rather than holding onto the entire
module.
