+++ src/main/kotlin/org/araqnid/kotlin/setawsssocredentials/childProcess/Spawn.kt
@@ -1,0 +1,7 @@
+@file:JsModule("node:child_process")
+
+package org.araqnid.kotlin.setawsssocredentials.childProcess
+
+import node.childProcess.SpawnOptions
+
+external fun spawn(command: String, args: Array<out String>, options: SpawnOptions): Subprocess
